Break RemoteDriver into multiple source files.
RemoteDriver was kind of a mess, and because I'll need to add more features
to it, was becoming harder and harder to change. This change breaks the file
into multiple sources, which doesn't necessarily makes it easier to understand,
but at least makes it easier to add new things and eventually do a propert
cleanup.
